Chapter{14,15,16}: Remove days in the singing predicate
In the logical application of the theorems presented in these
chapters, the day type would be the boolean type; in that case, we can
use Sings x denote whether x is true. The days also made necessary a
lot of bookkeeping in the proofs, which made the proofs pretty
unreadable.